Fix for Invisible Previews of SVGs
I'm not certain that this is the best solution for this, but when using SVGs that have no default height or width set they have a height and width of `0px` in the preview. With this modification they will always take up the full container still honoring the max pixel dimensions defined previously.